+++ R/postprocessing.R
@@ -277,7 +277,7 @@ doTimelapse <- function(timelapse){
 ##' @title Move pictures from unsorted to sorted folders
 ##'
 ##' @param timelapse data.frame. A data frame of a timelapse file. Note, this should follow the timelapse template that I typically use. You cannot specify both a timelapse file and a Dataorganize file using this function
-##' @param do data.frame. A data frame of a DataOrganze file. This can be either from the DataOrganize program or from the \code{\link{dataOrganize}} function in this package. You cannot specify both a timelapse and DataOrganize file using this function.
+##' @param do data.frame. A data frame of a DataOrganze file. This can be either from the DataOrganize program or from the \code{\link{doFolder}} function in this package. You cannot specify both a timelapse and DataOrganize file using this function.
 ##' @param in.dir String. The directory containing the root folder for the timelapse file. For example, if all your images were a folder called "images" which sits in an external drive, labelled "F:", then you would specify the in.dir as "F:".
 ##' @param out.dir String. The directory where you want to store the sorted images
 ##' @param create.dirs Logical. Should the function create the directories it needs?
@@ -288,7 +288,7 @@ doTimelapse <- function(timelapse){
 ##' @details When this function creates its folder structure, it uses the Individuals column in the Timelapse output.
 ##' For some "species" (e.g., ghost, human, bird, rodent), we do not sort these by individual, therefore the Individuals column is a 0.
 ##' These species get assigned a folder of 00 for their number of individuals. I do not know how this will affect quality control and the workflow down the line.
-##' Generally, it should not be an issue but could result in NA values in the Individuals column of the \code{\link{APFun_env}} or errors in the \code{\link{dataorganize}} functions in this package.
+##' Generally, it should not be an issue but could result in NA values in the Individuals column of the \code{\link{calculateEvents}} or errors in the \code{\link{doFolder}} functions in this package.
 ##' Once this has been tested, I will update this.
 ##'
 ##' This function can "sort" pictures using either a timelapse csv or a dataorganize text file.